add spec for tipping
rename helper_methods to feature_spec_helpers; add spec for tipping;
refactor change_selection_spec
